// types/typechain/src/contracts/libs/Events.sol/MinterEvent.ts
export interface CollateralDepositedEventObject {
  account: string;
  collateralAsset: string;
  amount: BigNumber;
}